refactor tests into smaller files
