+++ tests/responses/alarm/device_alert_no_alertstate.xml
