+++ tests/responses/alarm/device_alert_off.xml
