+++ tests/responses/alarm/device_alert_on.xml
