+++ tests/responses/alarm/device_magenta_smoke_alarm.xml
