+++ tests/responses/base/device_list.xml
