+++ tests/responses/base/device_no_devicelock_element.xml
