+++ tests/responses/base/device_not_present.xml
