+++ tests/responses/base/device_with_umlaut_in_name.xml
