+++ tests/responses/switch/device_list.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" ?>
+<devicelist version="1">
+    <device identifier="08761 0000434" id="17" functionbitmask="896" fwversion="03.33" manufacturer="AVM" productname="FRITZ!DECT 200">
+        <present>1</present>
+        <name>Steckdose</name>
+        <switch>
+            <state>1</state>
+            <mode>auto</mode>
+            <lock>0</lock>
+            <devicelock>0</devicelock>
+        </switch>
+        <powermeter>
+            <power>0</power>
+            <energy>707</energy>
+        </powermeter>
+        <temperature>
+            <celsius>285</celsius>
+            <offset>0</offset>
+        </temperature>
+    </device>
+</devicelist>
