+++ tests/responses/thermostat/device_hkr_fritzos_7.xml
