+++ tests/responses/thermostat/device_hkr_fw_03_50.xml
