+++ tests/responses/thermostat/device_hkr_fw_03_54.xml
