+++ tests/responses/thermostat/device_hkr_no_temp_values.xml
