+++ tests/responses/thermostat/device_hkr_state_manual.xml
