+++ tests/responses/thermostat/device_hkr_state_off.xml
