+++ tests/responses/thermostat/device_hkr_state_on.xml
