+++ tests/responses/thermostat/device_list_battery_low.xml
