+++ tests/responses/thermostat/device_list_battery_ok.xml
