+++ tests/test_fritzhome.py
@@ -3,16 +3,9 @@
 from nose.tools import eq_, raises
 from unittest.mock import MagicMock
 
+from .helper import Helper
+
 from pyfritzhome import Fritzhome, InvalidError, LoginError
-
-__responses = {}
-
-
-def response(file: str):
-    if file not in __responses:
-        with open("tests/responses/" + file + ".xml", "r") as file:
-            __responses[file] = file.read()
-    return __responses[file]
 
 
 class TestFritzhome(object):
@@ -24,16 +17,16 @@ class TestFritzhome(object):
     @raises(LoginError)
     def test_login_fail(self):
         self.mock.side_effect = [
-            response("login_rsp_without_valid_sid"),
-            response("login_rsp_without_valid_sid"),
+            Helper.response("login_rsp_without_valid_sid"),
+            Helper.response("login_rsp_without_valid_sid"),
         ]
 
         self.fritz.login()
 
     def test_login(self):
         self.mock.side_effect = [
-            response("login_rsp_without_valid_sid"),
-            response("login_rsp_with_valid_sid"),
+            Helper.response("login_rsp_without_valid_sid"),
+            Helper.response("login_rsp_with_valid_sid"),
         ]
 
         self.fritz.login()
@@ -79,9 +72,9 @@ class TestFritzhome(object):
 
     def test_get_device_element(self):
         self.mock.side_effect = [
-            response("device_list"),
-            response("device_list"),
-            response("device_list"),
+            Helper.response("base/device_list"),
+            Helper.response("base/device_list"),
+            Helper.response("base/device_list"),
         ]
 
         element = self.fritz.get_device_element("08761 0000434")
@@ -97,16 +90,17 @@ class TestFritzhome(object):
 
     def test_get_device_by_ain(self):
         self.mock.side_effect = [
-            response("device_list"),
-            response("device_list"),
+            Helper.response("base/device_list"),
+            Helper.response("base/device_list"),
         ]
 
+        self.fritz.update_devices()
         device = self.fritz.get_device_by_ain("08761 0000434")
         eq_(device.ain, "08761 0000434")
 
     def test_aha_get_devices(self):
         self.mock.side_effect = [
-            response("device_list"),
+            Helper.response("base/device_list"),
         ]
         self.fritz.update_devices()
 
@@ -142,10 +136,3 @@ class TestFritzhome(object):
             "http://10.0.0.1/webservices/homeautoswitch.lua",
             {"sid": None, "ain": "1", "switchcmd": "sethkrtsoll", "param": 254},
         )
-
-    def test_get_alert_state(self):
-        self.mock.side_effect = [
-            response("device_list"),
-        ]
-
-        eq_(self.fritz.get_alert_state("05333 0077045-1"), True)
